Move Ltl stuff to Cooked.Ltl
