+++ cooked-validators/cooked-validators.cabal
@@ -15,9 +15,9 @@ library
       Cooked
       Cooked.Attack
       Cooked.Currencies
+      Cooked.Ltl
       Cooked.MockChain
       Cooked.MockChain.Constraints
-      Cooked.MockChain.Ltl
       Cooked.MockChain.Monad
       Cooked.MockChain.Monad.Contract
       Cooked.MockChain.Monad.Direct
