+++ cooked-validators/tests/Cooked/AttackSpec.hs
@@ -12,8 +12,8 @@ module Cooked.AttackSpec (tests) where
 
 import Control.Monad
 import Cooked.Attack
+import Cooked.Ltl
 import Cooked.MockChain
-import Cooked.MockChain.Ltl
 import Cooked.Tx.Constraints
 import Data.Default
 import qualified Ledger as L hiding (singleton, validatorHash)
